Add udacity exercises repository (udacity/CVND_Exercises) :
https://github.com/udacity/CVND_Exercises.git

This repository contains code exercises and materials for Udacity's Computer Vision Nanodegree program. It consists of tutorial notebooks that demonstrate, or challenge you to complete, various computer vision applications and techniques. These notebooks depend on a number of software packages to run, and so, we suggest that you create a local environment with these dependencies by following the instructions below.